+++ common/src/main/java/tachyon/network/protocol/RPCMessage.java
@@ -22,7 +22,7 @@ import io.netty.buffer.ByteBuf;
 import io.netty.handler.codec.ByteToMessageDecoder;
 import io.netty.handler.codec.LengthFieldBasedFrameDecoder;
 
-import tachyon.network.protocol.buffer.DataBuffer;
+import tachyon.network.protocol.databuffer.DataBuffer;
 
 // This is the main base class for all RPC messages to the DataServer.
 // The message and type encoding scheme is adapted from the implementation found in the streaming
