+++ servers/src/test/java/tachyon/network/protocol/RPCBlockResponseTest.java
@@ -24,7 +24,7 @@ import org.junit.Test;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 
-import tachyon.network.protocol.buffer.DataByteBuffer;
+import tachyon.network.protocol.databuffer.DataByteBuffer;
 
 public class RPCBlockResponseTest {
   private static final long BLOCK_ID = 1;
